requirements.txt
python-chess>=1.9.0
numpy>=1.24.0
pygame>=2.5.0
scikit-learn>=1.3.0
tqdm>=4.65.0
